Update 20/03/2020
Addition of Database connection. RTs for 4 points was also include to be plot. Nonetheless only in those cases where this info is in the database it will be available.
This update creates a new version of Analyzer, since a new class was created to include the fields regarding database information.